Move token-groups to base theme directory
