+++ polaris-tokens/src/themes/base/breakpoints.ts
@@ -1,4 +1,4 @@
-import type {MetadataProperties} from '../types';
+import type {MetaTokenProperties} from '../types';
 
 // NOTE: Order is important here: smallest -> largest
 // Exporting as const means it will be typed as a Tuple instead of string[]
@@ -14,7 +14,7 @@ export type BreakpointsTokenGroup = {
 };
 
 export const breakpoints: {
-  [TokenName in BreakpointsTokenName]: MetadataProperties;
+  [TokenName in BreakpointsTokenName]: MetaTokenProperties;
 } = {
   'breakpoints-xs': {
     value: '0px',
