+++ polaris-tokens/src/themes/base/color.ts
@@ -1,6 +1,7 @@
-import type {MetadataProperties, Experimental} from '../types';
-import * as colors from '../colors';
-import * as colorsExperimental from '../colors-experimental';
+import type {Experimental} from '../../types';
+import type {MetaTokenProperties} from '../types';
+import * as colors from '../../colors';
+import * as colorsExperimental from '../../colors-experimental';
 
 export type ColorBackgroundAlias =
   | 'bg'
@@ -215,7 +216,7 @@ export type ColorTokenGroup = {
 };
 
 export const color: {
-  [TokenName in ColorTokenName]: MetadataProperties;
+  [TokenName in ColorTokenName]: MetaTokenProperties;
 } = {
   'color-bg-inverse': {
     value: colors.gray[900],
